moving session/transaction fixtures; adding dynamic scope to session/transaction
